Fix reduce view row collation with unicode equivalent keys
Previously, view reduce collation with keys relied on the keys in the
rows returned from the view shards to exactly match (=:=) the keys
specified in the args. However, in the case when there are multiple
rows which compare equal with the unicode collator, that may not
always be the case.

In that case when the rows are fetched from the row dict by key, they
should be matched using the same collation algorithm as the one used
on the view shards. Since, the collation module we use doesn't export
an equivalence function, we mimic it by using the `(not A < B) and
(not B < A)` logic. This relies on the subtle fact that all the things
we pass into the less/2 function have a defined comparison
order. Another option is to use the Erlang standard library
`string:equal(A, B, false, nfd)` function, which is more
direct. However, it could eventually hit a case when the Erlang
standard libraries NFD equivalence algorithm migth not agree with our
collation library's (libicu) idea of equivalence.